Flytter schemaer ut til egne filer
De er relativt små filer, men heller dette enn én stor schema-fil tror jeg
