+++ src/sider/08-vedlegg/form/BruttoNetto.tsx
@@ -2,13 +2,11 @@ import {DokumentasjonDtoType, LonnsInntektDto, type LonnsInput} from "../../../g
 import {useForm} from "react-hook-form";
 import React from "react";
 import {zodResolver} from "@hookform/resolvers/zod";
-import {z} from "zod";
 import {OpplysningBelopInput} from "./components/OpplysningBelopInput.tsx";
 import {useDokumentasjonTekster} from "../../../lib/hooks/dokumentasjon/useDokumentasjonTekster.ts";
-import {BruttoNettoFormSchema} from "./lib/formSchemas.ts";
 import {bruttoNettoFormToLonnsInput} from "./lib/formToInputMappers.ts";
+import {BruttoNettoFormSchema, BruttoNettoFormValues} from "./schema/bruttoNettoForm.ts";
 
-export type BruttoNettoFormValues = z.infer<typeof BruttoNettoFormSchema>;
 const BruttoNetto = ({
     opplysningstype,
     mutate,
